certs: reduce reload time for symlinks to 15 min
This commit reduces the certificate reload time
window (for symlinks) to 15 min. Before, certificates
would have been reloaded every 24h - which can be
an issue when certificates are valid for less than 24h.

Now, certificates mounted as symlinks will be reloaded
after 15 minutes - which is on the one side frequent enough
to avoid expiry issues and on the other side rare enough
to not cause any noticeable I/O load.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>