Add LDAP SRV Record lookup support
Adds `SRVRecordName` configuration parameter and validation for it. The
value can be `ldaps`, `ldap` or the special value `on`. This value is
combined with the `ServerAddr` parameter to lookup a DNS SRV record for
finding target LDAP servers.

The SRV record name format is always `_service._proto.domainname`. For LDAP
`proto` value here is always `tcp`. The `service` value is usually
`ldap`, and sometimes `ldaps`. When `SRVRecordName` is set to `ldap` or
`ldaps`, this value is substituted for `service`. The `domainname` value
is given by the `ServerAddr` parameter.

When a different SRV record name needs to be given, set `SRVRecordName`
to the special value `on`, and set the `ServerAddr` to the full SRV
record name, e.g. `_ldapish._tcpish.myldapserver.com`.