fix: bucket matching for bucket policy resource
This fixes some corner cases in bucket matching in bucket policy
resources.

Failing tests for bucket/policy resource in existing implementation are
added.

This happens because we cannot parse out a unique bucket pattern from a
resource string in all cases: when resource pattern is
`mybucket*a/myobject*`, it should match `mybucket/a/myobject`.

For this pattern, the buckets matched are:
   1. `mybucket` (e.g. resource value is "mybucket/a/myobject"),
   2. `mybucketXa`, where X is a string not containing a `/` (e.g.
   resource value is "mybucket-east-a/myobject"), and
   3. `mybucketY`, where Y is a string not containing a `/` (e.g.
   resource value that matches is "mybucket22/a/myobject")

To fix this a new MatchAsPatternPrefix function added to wildcard
package and bucket name matching code is updated.