Exit if license expired (#101)
* Exit if license expired

If the license has expired (and grace period also over where
applicable), then exit from the program.

To achieve this, the function `ValidateEnterpriseLicense` which is
called at application startup will also schedule the next license check
after successful validation.

If the license is currently not expired, schedule to check again just
after expiry. If it is expired but within grace period, check every day
(and print a warning message if it hasn't been renewed)

* Support setting license token in env MINIO_LICENSE

* Return license error through channel

and let the application log it in it's own way and exit.