Improvements in enterprise license validator (#103)
There could be specific flows where the caller doesn't want to be
notified when the license expires. In such cases, the caller can pass
the expiry channel as nil; in that case, the goroutine to schedule
future license validations should not be spawned.

In some cases, e.g., in the enterprise-console, the caller may want to explicitly pass 
the license token for validation. Accept it in the param LicenseToken, and when passed, 
do not try to read the token from the env variable or minio.license file.