Keep license token read from file
Currently if the license token is read from the license file, the
variable LicenseToken on the verifier remains empty, and caller cannot
get the token value without reading the file again.

Keeping the read token in LicenseToken variable will help the caller to
use it if required e.g. to automatically set the subnet license config
if not present.